+++ CHANGELOG.md
@@ -19,11 +19,26 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security
 
-## [0.24.1] - 2024-04-17
+## [0.25.1] - 2024-05-21
 
 ### Fixed
 
 - isolate settings context before first test
+
+## [0.25.0] - 2024-05-14
+
+### Added
+
+- add `precision` keyword to TemporalEntity constructor
+- add transform function for single wikidata organization to extracted organization
+
+### Changes
+
+- add tests for ldap.extract
+
+### Fixed
+
+- fix ldap.extract.get_merged_ids_by_email
 
 ## [0.24.0] - 2024-04-12